Moved common.nbt package to nbt as it doesn't require a platform dependent implementation.
