+++ src/main/java/com/wolfyscript/utilities/nbt/NBTTagConfigListPrimitive.java
@@ -1,0 +1,42 @@
+/*
+ *       WolfyUtilities, APIs and Utilities for Minecraft Spigot plugins
+ *                      Copyright (C) 2021  WolfyScript
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.wolfyscript.utilities.nbt;
+
+import com.fasterxml.jackson.annotation.JacksonInject;
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.wolfyscript.utilities.common.WolfyUtils;
+import java.util.List;
+
+public abstract class NBTTagConfigListPrimitive<VAL, T extends NBTTagConfigPrimitive<VAL>> extends NBTTagConfigList<T> {
+
+    @JsonCreator
+    NBTTagConfigListPrimitive(@JacksonInject WolfyUtils wolfyUtils, @JsonProperty("values") List<T> elements, Class<T> elementType) {
+        super(wolfyUtils, elements, elementType);
+    }
+
+    public NBTTagConfigListPrimitive(WolfyUtils wolfyUtils, NBTTagConfig parent, Class<T> elementType, List<T> values) {
+        super(wolfyUtils, parent, elementType, values);
+    }
+
+    public NBTTagConfigListPrimitive(NBTTagConfigList<T> other) {
+        super(other);
+    }
+
+}
